backend/requirements.txt
# Web 框架
fastapi==0.115.6
uvicorn[standard]==0.32.1

# 数据库
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# 配置
pydantic==2.10.3
pydantic-settings==2.6.1

# AI 框架
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
langgraph==0.2.60

# 加密
pycryptodome==3.21.0

# JWT
python-jose[cryptography]==3.3.0

# HTTP 客户端
httpx==0.27.2

# 图片处理
Pillow==11.0.0

# Redis（LangChain Memory + 任务缓存）
redis==5.2.1

# 定时任务
apscheduler==3.10.4

# MinIO 对象存储
minio==7.2.11

# 开发工具
python-dotenv==1.0.1
